WIP: Convert AnchorBoxes Layer to a numpy array
There's really no need that I can see for making this a Keras layer. Worse
is that there's a tile()-ing being done for each example in a batch which
means that this will take up much more space that it needs to.

Most importantly: this no longer functions. I'm not sure what was happening
in previous versions of Keras/Tensorflow, but this Model is specifically
agnostic to batch size at build time. This being the case, the tile() will
not work since it needs to know over how many examples it needs to tile.